feat: Add `—versionFilePathPrefix` option
Scenario: I have 2 React apps hosted at app.com/subdirectory1 and app.com/subdirectory2. I need the version.json for each app to be served from the correct subdirectory, but served from the normal build directory. In Vite this is known as the ‘base’ and now this option allows for specification of the correct subdirectory (e.g. —versionFilePathPrefix=subdirectory1/) so that the version.json path is output as `subdirectory1/version.json` in the HTML.